add OOP encapsulation and inheritance
