+++ teacher.rb
@@ -1,5 +1,6 @@
+require './person'
 class Teacher < Person
-  def initialize (specialization, classroom, name, parent_permission: 'true')
+  def initialize(specialization, name, parent_permission: 'true')
     super(age, name, parent_permission)
     @specialization = specialization
   end
